Move buffer into its own module
